Moving constants to defining files
