+++ vscode-dotnet-runtime-library/src/IDotnetAcquireContext.ts
@@ -3,7 +3,7 @@
  * Licensed under the MIT License. See License.txt in the project root for license information.
  * ------------------------------------------------------------------------------------------ */
 
-import { AcquireErrorConfiguration } from './Utils/ErrorConstants';
+import { AcquireErrorConfiguration } from './Utils/ErrorHandler';
 
 export interface IDotnetAcquireContext {
     version: string;
