+++ vscode-dotnet-runtime-library/src/IDotnetEnsureDependenciesContext.ts
@@ -3,7 +3,7 @@
  * Licensed under the MIT License. See License.txt in the project root for license information.
  * ------------------------------------------------------------------------------------------ */
 import * as cp from 'child_process';
-import { EnsureDependenciesErrorConfiguration } from './Utils/ErrorConstants';
+import { EnsureDependenciesErrorConfiguration } from './Utils/ErrorHandler';
 
 export interface IDotnetEnsureDependenciesContext {
     command: string;
