+++ vscode-dotnet-runtime-library/src/IDotnetUninstallContext.ts
@@ -3,7 +3,7 @@
  * Licensed under the MIT License. See License.txt in the project root for license information.
  * ------------------------------------------------------------------------------------------ */
 
-import { UninstallErrorConfiguration } from './Utils/ErrorConstants';
+import { UninstallErrorConfiguration } from './Utils/ErrorHandler';
 
 export interface IDotnetUninstallContext {
     errorConfiguration?: UninstallErrorConfiguration;
