--- vscode-dotnet-runtime-library/src/Utils/Configuration.ts
@@ -1,21 +1,0 @@
-/* --------------------------------------------------------------------------------------------
- * Copyright (c) Microsoft Corporation. All rights reserved.
- * Licensed under the MIT License. See License.txt in the project root for license information.
- * ------------------------------------------------------------------------------------------ */
-
-export const commandPrefix = 'dotnet'; // Prefix for commands
-
-export namespace commandKeys {
-    export const acquire = 'acquire';
-    export const uninstallAll = 'uninstallAll';
-    export const showAcquisitionLog = 'showAcquisitionLog';
-    export const ensureDotnetDependencies = 'ensureDotnetDependencies';
-    export const reportIssue = 'reportIssue';
-}
-
-export const configPrefix = 'dotnetAcquisitionExtension'; // Prefix for user settings
-
-export namespace configKeys {
-    export const installTimeoutValue = 'installTimeoutValue';
-    export const enableTelemetry = 'enableTelemetry';
-}
